+++ wumpus/wumpus-core/demo/TransformPath.hs
@@ -40,7 +40,7 @@ pic2 = pth `picOver` ch `picOver` xy_frame "rotate 30deg"
      pth  = mkBlackPath (rotate ang) pt
      ch   = rotate ang $ zcrosshair pt
      pt   = P2 70 10
-     ang  = d2r (30::Double)
+     ang  = d2r 30
 
 
 pic3 :: Picture
@@ -50,7 +50,7 @@ pic3 = pth `picOver` ch `picOver` xy_frame "rotateAbout (60,0) 30deg"
      ch   = rotateAbout ang pto $ zcrosshair pt
      pt   = P2 70 10
      pto  = P2 60 0 `asTypeOf` dpt
-     ang  = d2r (30::Double)
+     ang  = d2r 30
     
 
 pic4 :: Picture
